[REL] Migration product_image_url to 11.0
[FIX] change "openerp__.py" to manifest
[FIX] readapt code to set the image to pyhton 3
[FIX] change version code
