+++ product_image_url/README.rst
@@ -1,30 +1,74 @@
-Product Image From Web v9
-==========================
+.. |company| replace:: ADHOC SA
+
+.. |company_logo| image:: https://raw.githubusercontent.com/ingadhoc/maintainer-tools/master/resources/adhoc-logo.png
+   :alt: ADHOC SA
+   :target: https://www.adhoc.com.ar
+
+.. |icon| image:: https://raw.githubusercontent.com/ingadhoc/maintainer-tools/master/resources/adhoc-icon.png
+
+.. image:: https://img.shields.io/badge/license-AGPL--3-blue.png
+   :target: https://www.gnu.org/licenses/agpl
+   :alt: License: AGPL-3
+
+======================
+Product Image From Web
+======================
+
 This Cybrosys's module allows you to easily attach product image directly from the internet.
 You don't need to bother the product for an image.
 Just Google the product's name and identify & select it's photo. Copy the image URL and paste in image URL field in Odoo.
 This module will attach the corresponding photo in product form.
 
-Features
-========
-* Product Images from Internet.
-* Just copy the image address & paste it in "Image URL" field.
+Installation
+============
 
-Technical Notes
-===============
-Used Libraries:
+To install this module, you need to:
 
-* PIL
-* urllib2
-* requests
-* urllib2
-* base64
+#. Just install the module.
+
+Configuration
+=============
+
+#. Not Configuration needed.
+
+Usage
+=====
+
+#. Just copy the image address & paste it in "Image URL" field.
+
+
+.. image:: https://odoo-community.org/website/image/ir.attachment/5784_f2813bd/datas
+   :alt: Try me on Runbot
+   :target: http://runbot.adhoc.com.ar/
+
+Bug Tracker
+===========
+
+Bugs are tracked on `GitHub Issues
+<https://github.com/ingadhoc/product/issues>`_. In case of trouble, please
+check there if your issue has already been reported. If you spotted it first,
+help us smashing it by providing a detailed and welcomed feedback.
 
 Credits
 =======
+
 * Cybrosys Techno Solutions<https://www.cybrosys.com>
 
-Author
+Images
 ------
+
+* |company| |icon|
+
+Contributors
+------------
+
 Developer: Nilmar Shereef @ cybrosys, shereef@cybrosys.in
 
+Maintainer
+----------
+
+|company_logo|
+
+This module is maintained by the |company|.
+
+To contribute to this module, please visit https://www.adhoc.com.ar.
